V1.85(2014-11-6 18:51) 1.CurThreadId()方法返回值修正为uint64

+++ include/mdk/mapi.h
@@ -38,7 +38,7 @@ namespace mdk
 	//maxCpu可能的最大cpu数，如果检测出cpu数量大于此值，则表示检测有误，将返回defaultCpuNumber
 	//defaultCpuNumber默认cpu数量，当检测出错时，将认为cpu数量为此值
 	unsigned int GetCUPNumber( int maxCpu, int defaultCpuNumber );
-	unsigned int CurThreadId();//当前线程id
+	uint64 CurThreadId();//当前线程id
 	time_t mdk_Date();//返回0时0分0秒的当前日期
 	bool GetExeDir( char *exeDir, int size );//取得可执行程序位置
 }
